pyproject.toml
[project]
name = "langc-explore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community==0.3.1",
    "langchain-anthropic",
    "langchain-chroma",
    "langchain-huggingface",
    "sentence-transformers",
    "langgraph",
    "langsmith",
    "chromadb",
    "pypdf",
    "streamlit",
    "python-dotenv",
]
